% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-classes-dmi.R
\docType{class}
\name{dmi-class}
\alias{dmi-class}
\title{Data-Model Instance (per subject or per group)}
\description{
Data-Model Instance (per subject or per group)
}
\section{Slots}{

\describe{
\item{\code{model}}{model}

\item{\code{data}}{ANY}

\item{\code{node_1_index}}{ANY|NULL         # LBA/fastdm routing (OK NULL otherwise)}

\item{\code{is_positive_drift}}{ANY|NULL    # LBA/fastdm legacy flag/vector (NULL otherwise)}

\item{\code{q_matrix}}{matrix|NULL          # CDM: J x K}

\item{\code{prior_pi}}{numeric|NULL         # CDM: length 2^K, sums to 1}
}}

