% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-classes-model.R
\docType{class}
\name{model-class}
\alias{model-class}
\title{Base cognitive model (abstract)}
\description{
Base cognitive model (abstract)
}
\section{Slots}{

\describe{
\item{\code{parameter_map}}{list}

\item{\code{accumulators}}{character|NULL}

\item{\code{factors}}{list}

\item{\code{match_map}}{list|NULL}

\item{\code{constants}}{numeric|NULL}

\item{\code{cell_names}}{character}

\item{\code{parameter_x_condition_names}}{character}

\item{\code{model_boolean}}{ANY}

\item{\code{pnames}}{character}

\item{\code{npar}}{integer}
}}

