% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{BuildModel}
\alias{BuildModel}
\title{Build a model object}
\usage{
BuildModel(
  p_map = list(A = "1", B = "1", mean_v = "M", sd_v = "1", st0 = "1", t0 = "1"),
  accumulators = c("r1", "r2"),
  factors = list(S = c("s1", "s2")),
  match_map = list(M = list(s1 = "r1", s2 = "r2")),
  constants = c(sd_v = 1, st0 = 0),
  type = "lba",
  verbose = TRUE
)
}
\arguments{
\item{p_map}{descibes the association between the parameter and the
experimental factor.}

\item{accumulators}{specifies the response names and its levels.}

\item{factors}{specifies a list of factors and their levels/conditions.}

\item{match_map}{describes which the stimulus condition matches which response
level, resulting in a correct or an incorrect response.}

\item{constants}{the argument allows the user to decide which parameter is set to a
constant value.}

\item{type}{the model type defined in the package, "rd" or "lba".}

\item{verbose}{print design information}
}
\description{
The function performs a series of syntax checks to ensure the user enters
strings/values conforming the C++ internal setting.
}
\examples{
## A diffusion decision model
\dontrun{
model <- BuildModel(
    p_map = list(
        a = "1", v = "1", z = "1", d = "1", sz = "1", sv = "1",
        t0 = "1", st0 = "1"
    ),
    match_map = list(M = list(s1 = "r1", s2 = "r2")),
    factors = list(S = c("s1", "s2")),
    constants = c(st0 = 0, d = 0),
    accumulators = c("r1", "r2"),
    type = "rd"
)
}
\dontrun{
## A LBA model
model <- BuildModel(
    p_map = list(A = "1", B = "1", mean_v = "M", sd_v = "1", st0 = "1", t0 = "1"),
    match_map = list(M = list(s1 = "r1", s2 = "r2")),
    factors = list(S = c("s1", "s2")),
    constants = c(A = 0.75, mean_v.false = 1.5, sd_v = 1, st0 = 0),
    accumulators = c("r1", "r2"),
    type = "lba"
)
}

}
