% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{dmi-class}
\alias{dmi-class}
\title{The S4 class for the Data Model Instance}
\description{
The class is to represent a data-model instance, joinING a model object
with a data list.
}
\section{Slots}{

\describe{
\item{\code{model}}{A model object}

\item{\code{data}}{A data list or a data frame (for fitting parameters).}

\item{\code{node_1_index}}{a LBA specific parameter}

\item{\code{is_positive_drift}}{another LBA specific parameter}
}}

