% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{dmi-class}
\alias{dmi-class}
\title{The S4 class for the Data Model Instance}
\description{
The class is to represent a data-model instance, joining a model object
with a data list.
}
\section{Slots}{

\describe{
\item{\code{model}}{A model object}

\item{\code{data}}{A data list}

\item{\code{node_1_index}}{A LBA specific parameter, indicating the node 1 index}

\item{\code{is_positive_drift}}{Whether the drift rate is positive}
}}

