% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{model-class}
\alias{model-class}
\title{The S4 class for a model object.}
\description{
model class represents how the model parameter associates with the
experimental factor.
}
\section{Slots}{

\describe{
\item{\code{parameter_map}}{indicates how model parameters associate with  experimental factors.}

\item{\code{accumulators}}{A character vector representing the name of the model accumulator (in the case of the accumulator model).}

\item{\code{factors}}{A list of factors defined in a design.}

\item{\code{match_map}}{defines what constitutes a correct and an incorrect response.}

\item{\code{constants}}{A list that informs what parameters the user sets to a fixed value.}

\item{\code{cell_names}}{A character vector representing the name of the condition.}

\item{\code{parameter_x_condition_names}}{A character vector representing the association of the design and the model.}

\item{\code{model_boolean}}{A 3-D TRUE/FALSE array for allocating parameters to conditions.}

\item{\code{pnames}}{a string vector specifying the names of the free paraemters}

\item{\code{npar}}{the number of the free parameter}

\item{\code{type}}{a string sepcifying which model to fit}
}}

